Split sidebar field panels into standalone components
Extract the connector field into a dedicated ConnectorField component
with its own ConnectorSettingsPopover, and split the merged
UserPickerFieldPanel into separate AssigneesField (editable) and
ParticipantsField (read-only) components so each only pulls in the
hooks and state it actually needs.
